requirements.txt
fastapi==0.104.0
pydantic==2.4.2
uvicorn==0.23.2
strawberry-graphql==0.209.7
